Move node-authorizer config files to /etc/srv/kubernetes/node-authorizer
/srv isn't writeable on Google's COS or other OSes.  Instead we write
to /etc/srv/kubernetes; that should be writeable on all OSes.  This
means we don't need to template the config dir.

Also starting to split files by component, so we write to
/etc/srv/kubernetes/node-authorizer; this means each component can
just mount the directory and get (just) the files it needs.  Should be
simpler and more secure.

Finally, change the extensions that the CreatePrivateKeyPairTask
creates, to better match what we expect elsewhere (.crt and .key).